float to int

+++ scripts/query_yle.py
@@ -45,7 +45,7 @@ def _parse_arguments():
     parser.add_argument('-lang', '--language',
                         help="language to search (fi=YLE uutiset,sv=Svenska YLE)", default="fi")
     parser.add_argument(
-        '-d', '--delay', help="number of seconds to wait between consecutive requests", default=1.0, type=float)
+        '-d', '--delay', help="number of seconds to wait between consecutive requests", default=1, type=int)
     parser.add_argument('--quiet', default=False,
                         action='store_true', help="Log only errors")
     return parser.parse_args()
